requirements.txt
flask==2.0.3
flask-sqlalchemy==2.5.1
mysqlclient==2.1.0
python-dotenv==0.19.2
requests==2.27.1
sqlalchemy-utils==0.38.2
six==1.16.0